Fixed delete feature user in the assigned part

See merge request acvt719/task-management-tool!46

+++ app.py
@@ -187,48 +187,54 @@ def update_profile(user_email):
 # Delete the user
 @app.route('/dashboard/<user_email>/delete_user', methods=['POST'])
 def delete_user(user_email):
-    # Delete all the project's of the user including everything related to it
+
+    # 1. Do not forget to delete the user from every assigned task
+    fullname = get_user_fullname(user_email)
+    mongo.db.task_table.update({'assign_to': fullname}, {'$unset': {'assign_to': 1}}, multi=True)
+
+    # 2. Do not forget to delete the user from the assigned_user table
+    mongo.db.assigned_table.remove({'email': user_email})
+
+    # 3. Delete all the project's of the user including everything related to it
     project_collection_list = list(mongo.db.project_table.find({'project_creator_email': user_email}, {'_id': 1}))
 
-    for x in project_collection_list:
-        # we want to take only the substring which is inside the parenthesis (the id)
-        # but it will be easier to convert the whole dict to a string first
-        project_id_string = str(x)
-        # gives '6060530cf082cd618caaad54' and then 6060530cf082cd618caaad54
-        project_id = project_id_string[project_id_string.find("(") + 1:project_id_string.find(")")].replace("'", "")
-        project_id = ObjectId(project_id)
+    if project_collection_list:
+        for x in project_collection_list:
+            # we want to take only the substring which is inside the parenthesis (the id)
+            # but it will be easier to convert the whole dict to a string first
+            project_id_string = str(x)
+            # gives '6060530cf082cd618caaad54' and then 6060530cf082cd618caaad54
+            project_id = project_id_string[project_id_string.find("(") + 1:project_id_string.find(")")].replace("'", "")
+            project_id = ObjectId(project_id)
 
-        # 1. Do not forget to delete the assigned users of the project's tasks
-        mongo.db.assigned_table.remove({"project_id": project_id})  #
+            # 3A. Delete the uploaded files of the project
+            upload_results = mongo.db.upload_table.find(
+                {
+                    "project_id": project_id})  # find the collection of the files which we want to delete based on the project_id
 
-        # 2. Delete the uploaded files of the project
-        upload_results = mongo.db.upload_table.find(
-            {
-                "project_id": project_id})  # find the collection of the files which we want to delete based on the project_id
+            for file_document in upload_results:  # iterate all the files which are included in the project that we want to delete
 
-        for file_document in upload_results:  # iterate all the files which are included in the project that we want to delete
+                file_name = file_document[
+                    'filename']  # keep the name of the file that we want to delete which exist in the upload_table
+                fs_result = mongo.db.fs.files.find_one({
+                    "filename": file_name})  # use the above filename in order to collect the data of the file which we want to delete and also exist in the fs.file table
+                fs_id = fs_result[
+                    '_id']  # keep the id of the file that we want to delete which it is also exist in the fs.file table
 
-            file_name = file_document[
-                'filename']  # keep the name of the file that we want to delete which exist in the upload_table
-            fs_result = mongo.db.fs.files.find_one({
-                "filename": file_name})  # use the above filename in order to collect the data of the file which we want to delete and also exist in the fs.file table
-            fs_id = fs_result[
-                '_id']  # keep the id of the file that we want to delete which it is also exist in the fs.file table
+                fs.delete(fs_id)  # delete the file from the fs.files and fs.chunks table
+                mongo.db.upload_table.remove({'project_id': project_id})  # delete the file from the upload_table
 
-            fs.delete(fs_id)  # delete the file from the fs.files and fs.chunks table
-            mongo.db.upload_table.remove({'project_id': project_id})  # delete the file from the upload_table
+            # 3B. Delete the tasks of the user's project
+            mongo.db.task_table.remove({'project_id': project_id})
 
-        # 3. Delete the tasks of the project
-        mongo.db.task_table.remove({'project_id': project_id})  # Delete the project's tasks from the database
+            # 3C. Delete all the user's projects
+            mongo.db.project_table.remove({'_id': project_id})
 
-        # 4. Delete the project
-        mongo.db.project_table.remove({'_id': project_id})  # Delete the project from the database
-
-    # 5: Delete the user
+    # 4: Finally delete the user
     mongo.db.user_table.remove({'email': user_email})
 
-    #  return redirect(url_for('view_profile', user_email=user_email))
     return render_template('dashboard/alerts/alertDeleteProfile.html', user_email=user_email)
+
 
 # Get the fullname of the user
 def get_user_fullname(user_email):
@@ -758,7 +764,8 @@ def download_file(user_email, project_id, filename):
                 output.write(output_data)
                 output.close()
 
-                return render_template('dashboard/alerts/alertUploadFile.html', user_email=user_email, project_id=project_id)
+                return render_template('dashboard/alerts/alertUploadFile.html', user_email=user_email,
+                                       project_id=project_id)
 
             else:
                 return mongo.send_file(filename)
